+++ src/components/index.tsx
@@ -4,3 +4,5 @@ export { default as Footer } from "./Footer/Footer";
 export { default as Layout } from "./Layout/Layout";
 export { default as HeaderLink } from "./Header/HeaderLink/HeaderLink";
 export { default as MobileHeaderLink } from "./Header/MobileHeaderLink/MobileHeaderLink";
+export { default as NavList } from "./Header/NavList/NavList";
+export { default as MobileNavList } from "./Header/MobileNavList/MobileNavList";
